pyproject.toml
[project]
name = "postergeist"
version = "0.1.0"
description = "Academic poster generator from Markdown to HTML"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "mistune>=3.0",
    "pyyaml>=6.0",
    "qrcode[pil]>=7.0",
    "watchdog>=4.0",
    "pillow>=10.0",
    "playwright>=1.40",
]

[project.scripts]
postergeist = "postergeist.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/postergeist"]
